Move legacy components to old_components
