+++ frontend/src/apps/users/containers/password_reset_confirm.js
@@ -4,9 +4,9 @@ import {Form} from 'react-form';
 import autoBind from 'react-autobind';
 import cx from 'classnames';
 
-import {TextInput} from 'common/components/inputs';
-import {Button} from 'common/components/buttons';
-import {DoubleBounceSpinner} from 'common/components/spinners';
+import {TextInput} from 'common/old_components/inputs';
+import {Button} from 'common/old_components/buttons';
+import {DoubleBounceSpinner} from 'common/old_components/spinners';
 import {getQueryParams} from 'helpers/utils';
 import {post} from 'helpers/rest';
