+++ frontend/src/common/old_components/spinners/index.js
