+++ webpack-stats-prod.json
@@ -1,1 +1,1 @@
-{"status":"done","chunks":{"commons":[{"name":"commons-bundle.df9563f0cbd52a398de5.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/commons-bundle.df9563f0cbd52a398de5.js"}],"landing":[{"name":"landing-bundle.df9563f0cbd52a398de5.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/landing-bundle.df9563f0cbd52a398de5.js"}],"login":[{"name":"login-bundle.df9563f0cbd52a398de5.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/login-bundle.df9563f0cbd52a398de5.js"}],"password_reset":[{"name":"password_reset-bundle.df9563f0cbd52a398de5.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/password_reset-bundle.df9563f0cbd52a398de5.js"}],"password_reset_confirm":[{"name":"password_reset_confirm-bundle.df9563f0cbd52a398de5.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/password_reset_confirm-bundle.df9563f0cbd52a398de5.js"}],"signup":[{"name":"signup-bundle.df9563f0cbd52a398de5.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/signup-bundle.df9563f0cbd52a398de5.js"}]}}
+{"status":"done","chunks":{"commons":[{"name":"commons-bundle.c1594526233ef386731e.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/commons-bundle.c1594526233ef386731e.js"}],"landing":[{"name":"landing-bundle.c1594526233ef386731e.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/landing-bundle.c1594526233ef386731e.js"}],"login":[{"name":"login-bundle.c1594526233ef386731e.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/login-bundle.c1594526233ef386731e.js"}],"password_reset":[{"name":"password_reset-bundle.c1594526233ef386731e.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/password_reset-bundle.c1594526233ef386731e.js"}],"password_reset_confirm":[{"name":"password_reset_confirm-bundle.c1594526233ef386731e.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/password_reset_confirm-bundle.c1594526233ef386731e.js"}],"signup":[{"name":"signup-bundle.c1594526233ef386731e.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/build/webpack_bundles/signup-bundle.c1594526233ef386731e.js"}]}}
