+++ webpack-stats.json
@@ -1,1 +1,1 @@
-{"status":"done","chunks":{"commons":[{"name":"commons-bundle.59e5ed4ce0f646e378f4.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/commons-bundle.59e5ed4ce0f646e378f4.js"}],"landing":[{"name":"landing-bundle.59e5ed4ce0f646e378f4.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/landing-bundle.59e5ed4ce0f646e378f4.js"}],"login":[{"name":"login-bundle.59e5ed4ce0f646e378f4.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/login-bundle.59e5ed4ce0f646e378f4.js"}],"password_reset":[{"name":"password_reset-bundle.59e5ed4ce0f646e378f4.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/password_reset-bundle.59e5ed4ce0f646e378f4.js"}],"password_reset_confirm":[{"name":"password_reset_confirm-bundle.59e5ed4ce0f646e378f4.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/password_reset_confirm-bundle.59e5ed4ce0f646e378f4.js"}],"signup":[{"name":"signup-bundle.59e5ed4ce0f646e378f4.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/signup-bundle.59e5ed4ce0f646e378f4.js"}]}}
+{"status":"done","chunks":{"commons":[{"name":"commons-bundle.7781d7011a12661581e3.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/commons-bundle.7781d7011a12661581e3.js"}],"landing":[{"name":"landing-bundle.7781d7011a12661581e3.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/landing-bundle.7781d7011a12661581e3.js"}],"login":[{"name":"login-bundle.7781d7011a12661581e3.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/login-bundle.7781d7011a12661581e3.js"}],"password_reset":[{"name":"password_reset-bundle.7781d7011a12661581e3.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/password_reset-bundle.7781d7011a12661581e3.js"}],"password_reset_confirm":[{"name":"password_reset_confirm-bundle.7781d7011a12661581e3.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/password_reset_confirm-bundle.7781d7011a12661581e3.js"}],"signup":[{"name":"signup-bundle.7781d7011a12661581e3.js","path":"/Users/valeriikovalchuk/projects/holoapollo/frontend/dist/webpack_bundles/signup-bundle.7781d7011a12661581e3.js"}]}}
